refactor: update import paths to use relative paths from core module
