+++ packages/vscode/src/vshost.ts
@@ -2,18 +2,18 @@
 // Licensed under the MIT License.
 
 import * as vscode from "vscode";
-import { createNodePath, ElementOrArray } from "@genaiscript/core";
+import { createNodePath } from "../../core/src/path";
 import { TerminalServerManager } from "./servermanager";
 import { Uri } from "vscode";
 import { ExtensionState } from "./state";
 import { Utils } from "vscode-uri";
-import { filterGitIgnore } from "@genaiscript/core";
-import { setHost, Host } from "@genaiscript/core";
-import { TraceOptions } from "@genaiscript/core";
-import { arrayify } from "@genaiscript/core";
 import { uniq } from "es-toolkit";
-import { CancellationOptions } from "@genaiscript/core";
-import { LanguageModelConfiguration, LogLevel } from "@genaiscript/core";
+import { Host, setHost } from "../../core/src/host";
+import { arrayify } from "../../core/src/cleaners";
+import { filterGitIgnore } from "../../core/src/gitignore";
+import { CancellationOptions } from "../../core/src/cancellation";
+import { TraceOptions } from "../../core/src/trace";
+import { LanguageModelConfiguration, LogLevel } from "../../core/src/server/messages";
 
 export class VSCodeHost extends EventTarget implements Host {
   userState: any = {};
